Split internal trait
