+++ packages/framework/src/Foundation/Concerns/ManagesHydeKernel.php
@@ -22,53 +22,6 @@ use function is_subclass_of;
  */
 trait ManagesHydeKernel
 {
-    /** @var bool Is the Kernel currently booting? */
-    protected bool $booting = false;
-
-    /** @var bool Is the Kernel ready to be booted? */
-    protected bool $canBoot = false;
-
-    public function isBooted(): bool
-    {
-        return $this->booted;
-    }
-
-    /** @internal */
-    public function readyToBoot(): void
-    {
-        $this->canBoot = true;
-    }
-
-    public function boot(): void
-    {
-        if (! $this->canBoot) {
-            throw new BadMethodCallException('The HydeKernel cannot be booted yet.');
-        }
-
-        if ($this->isBooted()) {
-            return;
-        }
-
-        $this->booting = true;
-
-        $this->files = FileCollection::boot($this);
-        $this->pages = PageCollection::boot($this);
-        $this->routes = RouteCollection::boot($this);
-
-        $this->booting = false;
-        $this->booted = true;
-    }
-
-    /** @internal Reboot the kernel - useful for resetting the application during testing */
-    public static function reboot(): void
-    {
-        $kernel = static::getInstance();
-
-        $kernel->files = FileCollection::boot($kernel);
-        $kernel->pages = PageCollection::boot($kernel);
-        $kernel->routes = RouteCollection::boot($kernel);
-    }
-
     public static function getInstance(): HydeKernel
     {
         return static::$instance;
